move one time scripts to their own file
